+++ CODEFORCES/1000/1836B.cpp
@@ -1,0 +1,11 @@
+#include<bits/stdc++.h>
+using namespace std;
+long long T,n,k,g,l,r,mid,ans;
+int main(){
+  cin>>T;
+  while(T--){
+    cin>>n>>k>>g;
+    ans=min(k,((g-1)>>1)*n/g);
+    cout<<(long long)(ans)*g<<endl;
+  }
+}